View symptom record functionality
Same issue as with view irritant.  After adding record and hitting back button record is not on recyclerview.  However navigating between fragments causes it to populate.  Likely because population is in onCreate.  Potential fix would be to auto navigate back to mainActivity.fragment upon record added or populate onResume or some such thing.