update symptom record functional
The update symptom record is now functional. Issue of selected time not saving persists into update record as well.  Needs resolved.  Need to copy over to irritants.  For design purposes it would be best a button or indicator of some sort to inform that clicking on card will allow editing of record.  Would be good to have indicator of create or edit on add record activity as well.